add support for CUB-dataset
Add support to run the CUB dataset, which have not been touched for a while. Now there are utils for reading and saving on a file-structure pretty similar to Shapes. Evaluation and hyperparameters are also supported. This is made as separate function and files, even though they could be put together pretty easily. Ther reason to not make the files more generic is that this way leads to faster development, and change might occour that makes them better to separate later.