update formatting.
Update formatting in all files. The formatter used at the end of the thesis got depricated, and it was not fixed when doing the last changes for the thesis. Therefore, many formatting errors where present. Update them with flake8. Also removed two functions in plotting.py that were never used and had bugs.